tests/requirements.txt
selenium>=4.35.0
pytest>=8.0.0
pytest-html>=4.0.0
requests>=2.25.0
Appium-Python-Client>=3.1.0
pytest-xdist>=3.3.0
python-dotenv>=1.0.0